% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minzar.R
\name{minzar_test}
\alias{minzar_test}
\title{Mincer-Zarnowitz Test}
\usage{
minzar_test(actual, forecast, ...)
}
\arguments{
\item{actual}{a vector representing the actual values of a series.}

\item{forecast}{a vector representing the forecasted values of the series.}

\item{...}{additional arguments passed to \code{\link[car]{linearHypothesis}},
except the \dQuote{test} argument which is fixed to Chisq.}
}
\value{
An object of class \dQuote{tstest.minzar} which has a print and
as_flextable method.
}
\description{
The forecast unbiasedness test of Mincer and Zarnowitz (1969).
}
\examples{
data(arma_forecast)
test <- minzar_test(arma_forecast$actual, arma_forecast$forecast)
test

}
\references{
\insertRef{Mincer1969}{tstests}
}
