% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkowitz.R, R/dac.R, R/gmm.R, R/hongli.R,
%   R/minzar.R, R/nyblom.R, R/shortfall_de.R, R/signbias.R, R/var.R, R/var_cp.R
\name{print.tstest.berkowitz}
\alias{print.tstest.berkowitz}
\alias{print.tstest}
\alias{print.tstest.dac}
\alias{print.tstest.gmm}
\alias{print.tstest.hongli}
\alias{print.tstest.minzar}
\alias{print.tstest.nyblom}
\alias{print.tstest.shortfall_de}
\alias{print.tstest.signbias}
\alias{print.tstest.vares}
\alias{print.tstest.var_cp}
\title{Test Print method}
\usage{
\method{print}{tstest.berkowitz}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  ...
)

\method{print}{tstest.dac}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  ...
)

\method{print}{tstest.gmm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  collapse = TRUE,
  ...
)

\method{print}{tstest.hongli}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  ...
)

\method{print}{tstest.minzar}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  ...
)

\method{print}{tstest.nyblom}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  ...
)

\method{print}{tstest.shortfall_de}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  ...
)

\method{print}{tstest.signbias}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  ...
)

\method{print}{tstest.vares}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  ...
)

\method{print}{tstest.var_cp}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  include.decision = FALSE,
  ...
)
}
\arguments{
\item{x}{an object inheriting class \dQuote{tstest.test}.}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{signif.stars}{logical. If TRUE, \sQuote{significance stars} are printed.}

\item{include.decision}{prints out whether to reject the NULL at the 5\% level
of significance.}

\item{...}{not currently used.}

\item{collapse}{collapses the results for multiple lags to just report the
joint test.}
}
\value{
Invisibly returns the original object.
}
\description{
Print method for objects inheriting class \dQuote{tstest}
}
