% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spy}
\alias{spy}
\title{SPY ETF Adjusted Close}
\format{
\subsection{\code{spy}}{

An xts vector with 7597 observations spanning the period 1993-01-29 / 2023-03-30
from Yahoo Finance.
}
}
\usage{
spy
}
\description{
The adjusted closing price of the SPY ETF.
}
\keyword{datasets}
