% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstests-package.R
\docType{package}
\name{tstests-package}
\alias{tstests}
\alias{tstests-package}
\title{tstests: Tests and Metrics for timeseries}
\description{
Tests and metrics from timeseries analysis.
}
\keyword{internal}
