% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstests-package.R
\docType{package}
\name{tstests-package}
\alias{tstests}
\alias{tstests-package}
\title{tstests: Time Series Model and Forecast Tests}
\description{
Goodness of Fit and Forecast Evaluation Tests for timeseries models.
}
\keyword{internal}
