% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstests-package.R
\docType{package}
\name{tstests-package}
\alias{tstests}
\alias{tstests-package}
\title{tstests: Time Series Goodness of Fit and Forecast Evaluation Tests}
\description{
Goodness of Fit and Forecast Evaluation Tests for timeseries models. Includes, among others, the Generalized Method of Moments (GMM) Orthogonality Test of Hansen (1982), the Nyblom (1989) parameter constancy test, the sign-bias test of Engle and Ng (1993), and a range of tests for value at risk and expected shortfall evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.nopredict.com/packages/tstests}
  \item \url{https://github.com/tsmodels/tstests}
}

}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com} (\href{https://orcid.org/0009-0000-9308-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
